+++ src/Reed/Classes/Article/WebViewController.swift
@@ -63,7 +63,7 @@ final class WebViewController: UIViewController {
         NotificationCenter.default.addObserver(
             self,
             selector: #selector(self.avatarDidBecomeAvailable(_:)),
-            name: .AvatarDidBecomeAvailable,
+            name: .avatarDidBecomeAvailable,
             object: nil
         )
         NotificationCenter.default.addObserver(